sales module updated

+++ src/main/java/com/example/minierp/interfaces/rest/product/CreateProductRequest.java
@@ -2,10 +2,12 @@ package com.example.minierp.interfaces.rest.product;
 
 import jakarta.validation.constraints.*;
 
+import java.math.BigDecimal;
+
 public record CreateProductRequest(
         @NotBlank String  name,
         @NotBlank String sku,
-        @NotNull @Positive double price,
+        @NotNull @Positive BigDecimal price,
         @NotNull @Min(0) int quantity
 ) {
 }
