+++ src/main/java/com/example/minierp/interfaces/rest/reports/OrderReportResponse.java
@@ -2,6 +2,7 @@ package com.example.minierp.interfaces.rest.reports;
 
 import com.example.minierp.domain.sales.Order;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -11,7 +12,7 @@ public record OrderReportResponse(
         String status,
         List<Item> items
 ) {
-    public record Item(String productName, Integer quantity, Double price) {}
+    public record Item(String productName, Integer quantity, BigDecimal price) {}
 
     public static OrderReportResponse from(Order order) {
         return new OrderReportResponse(
